+++ bestmobabot/settings.py
@@ -21,6 +21,7 @@ class ArenaSettings(BaseModel):
     max_pages: int = 15  # maximal number of pages during enemy search
     max_grand_pages: int = 15  # maximal number of pages during grand enemy search
     hyper_params: Optional[dict] = None  # hyper-parameters of the predictive model
+    randomize_grand_defenders: bool = False
 
     # noinspection PyMethodParameters
     @validator('early_stop')