+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# `latest`
+
+* **Option to randomize grand arena defenders**
+
 # `2.0`
 
 **Breaking change.** As I keep adding more and more configuration options, I decided to move them out from the command line and environment variables to a separate configuration file. If you upgraded to this version, make sure you properly move your settings.